% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_cardio.R
\name{model_cardio}
\alias{model_cardio}
\title{Probabilidad de Enferemdades Cardio.}
\usage{
model_cardio(age, sex, thalach)
}
\arguments{
\item{age}{Ingresar edad del paciente "Female,Male"}

\item{sex}{Genero del paciente}

\item{thalach}{Presción sanginea del paciente}
}
\value{

}
\description{
Probabilidad de Enferemdades Cardio.
}
\examples{
model_cardio(30,"Female",100)

}
